pyproject.toml
[project]
name = "tanwei-bot"
version = "0.1.0"
description = "配置驱动的多 Agent 信息监控与推送工具"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"

dependencies = [
    "pyyaml>=6.0",
    "feedparser>=6.0",
    "httpx>=0.27",
    "click>=8.0",
    "schedule>=1.2",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[project.scripts]
tanwei-bot = "tanwei.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
